requirements.txt
Office365-REST-Python-Client
python-dotenv
pywin32
pandas
shortuuid
openpyxl